refactor(oidc): extract requests in another file (#931)
