+++ packages/vanilla-demo/public/.gitignore
@@ -1,0 +1,1 @@
+OidcServiceWorker.js
